Add remark command into app
The application is unable to take in remarks, and
only allows a set of variables to be defined.

This may cause the user to not remember each
contact precisely, such as forgetting who John Doe
is to them, as an example.

Add the ability to add a remark to a contact,
that will appear on the contact list.

This will allow the users to add custom remarks to
their contact of their choosing, allowing for
more customizability.

No further references outside of academic guide.