Fix Date Formatting
Users can enter an invalid date, and no error message would show up.

This makes it confusing as they are left wondering why their commands
is invalid.

Fix the date formatting to throw an error message on the output box
for the users to see by ensuring that commands are using the parser
method.

This will allow users to see the error.